+++ lib/rubocop/cop/style/space_inside_string_interpolation.rb
@@ -40,7 +40,7 @@ module RuboCop
         private
 
         def autocorrect(node)
-          new_source = (style == :no_space) ? node.source : " #{node.source} "
+          new_source = style == :no_space ? node.source : " #{node.source} "
           lambda do |corrector|
             corrector.replace(range_with_surrounding_space(node.source_range),
                               new_source)